Add runnable example
Creation of own file for runnable example which was previously inside Cit.py file.
 Cit.py should be just for computing combinations.
