+++ setup.py
@@ -1,0 +1,7 @@
+from setuptools import setup, find_packages
+
+setup(name='cit-ng',
+      description='Varianter with combinatorial capabilities',
+      packages=find_packages(),
+      include_package_data=True,
+      )
